refactor: My 폴더 -> Main 이동
